Shift sort logic to MyVideosView.jsx
